// Qanat.Web/src/app/shared/generated/model/allocation-plan-period-simple-dto.ts
export class AllocationPlanPeriodSimpleDto { 
    AllocationPlanPeriodID?: number;
    AllocationPlanID?: number;
    AllocationPeriodName?: string;
    AllocationAcreFeetPerAcre?: number;
    NumberOfYears?: number;
    StartYear?: number;
    EnableCarryOver?: boolean;
    CarryOverNumberOfYears?: number;
    CarryOverDepreciationRate?: number;
    EnableBorrowForward?: boolean;
    BorrowForwardNumberOfYears?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}