// Qanat.Web/src/app/shared/generated/model/allocation-plan-period-upsert-dto.ts
export class AllocationPlanPeriodUpsertDto { 
    AllocationPlanID: number;
    AllocationPlanPeriodID?: number;
    AllocationPeriodName: string;
    AllocationAcreFeetPerAcre: number;
    NumberOfYears: number;
    StartYear: number;
    EnableCarryOver: boolean;
    CarryOverNumberOfYears?: number | null;
    CarryOverDepreciationRate?: number | null;
    EnableBorrowForward: boolean;
    BorrowForwardNumberOfYears?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}